+++ src/gui/keywordWidgets/CMakeLists.txt
@@ -1,132 +1,119 @@
-# Don't forget to include output directory, otherwise the UI file won't be wrapped!
-
-# Meta-Objects
-set(keywordWidgets_MOC_HDRS
-    atomTypeVector.h
-    configuration.h
-    configurationVector.h
-    dialog.h
-    dropDown.h
-    expressionVariableVector.h
-    fileAndFormat.h
-    function1D.h
-    interactionPotential.h
-    isotopologueSet.h
-    layer.h
-    module.h
-    moduleVector.h
-    node.h
-    nodeAndInteger.h
-    nodeValue.h
-    nodeValueEnumOptions.h
-    nodeVector.h
-    range.h
-    species.h
-    speciesSite.h
-    speciesSiteVector.h
-    speciesVector.h
-    vec3Double.h
-    vec3Integer.h
-    vec3NodeValue.h
-    weightedModuleVector.h
-    rangeVector.h
-    bool.hui
-    double.hui
-    dropWidget.hui
-    enumOptions.hui
-    integer.hui
-    optionalDouble.hui
-    optionalInt.hui
-    stdString.hui
-    widget.hui
-    widgetGroupHeader.h
-)
-qt6_wrap_cpp(keywordWidgets_MOC_SRCS ${keywordWidgets_MOC_HDRS})
-
-# User Interface Files
-set(keywordWidgets_UIS
-    atomTypeVector.ui
-    configuration.ui
-    configurationVector.ui
-    dialog.ui
-    dropDown.ui
-    expressionVariableVector.ui
-    fileAndFormat.ui
-    function1D.ui
-    interactionPotential.ui
-    isotopologueSet.ui
-    layer.ui
-    module.ui
-    moduleVector.ui
-    node.ui
-    nodeAndInteger.ui
-    nodeValue.ui
-    nodeValueEnumOptions.ui
-    nodeVector.ui
-    range.ui
-    rangeVector.ui
-    species.ui
-    speciesSite.ui
-    speciesSiteVector.ui
-    speciesVector.ui
-    vec3Double.ui
-    vec3Integer.ui
-    vec3NodeValue.ui
-    weightedModuleVector.ui
-    widgetGroupHeader.ui
-)
-qt6_wrap_ui(keywordWidgets_UIS_H ${keywordWidgets_UIS})
-
-# Source files (not Qt UI files or custom .h [uih])
-set(keywordWidgets_SRCS
-    base.cpp
-    atomTypeVectorFuncs.cpp
-    boolFuncs.cpp
-    configurationFuncs.cpp
-    configurationVectorFuncs.cpp
-    dialogFuncs.cpp
-    doubleFuncs.cpp
-    dropDownFuncs.cpp
-    dropWidgetFuncs.cpp
-    enumOptionsFuncs.cpp
-    expressionVariableVectorFuncs.cpp
-    fileAndFormatFuncs.cpp
-    function1DFuncs.cpp
-    integerFuncs.cpp
-    interactionPotentialFuncs.cpp
-    isotopologueSetFuncs.cpp
-    layerFuncs.cpp
-    moduleFuncs.cpp
-    moduleVectorFuncs.cpp
-    nodeFuncs.cpp
-    nodeAndIntegerFuncs.cpp
-    nodeValueFuncs.cpp
-    nodeValueEnumOptionsFuncs.cpp
-    nodeVectorFuncs.cpp
-    optionalDoubleFuncs.cpp
-    optionalIntFuncs.cpp
-    producers.cpp
-    rangeFuncs.cpp
-    rangeVectorFuncs.cpp
-    speciesFuncs.cpp
-    speciesSiteFuncs.cpp
-    speciesSiteVectorFuncs.cpp
-    speciesVectorFuncs.cpp
-    stdStringFuncs.cpp
-    vec3DoubleFuncs.cpp
-    vec3IntegerFuncs.cpp
-    vec3Labels.cpp
-    vec3NodeValueFuncs.cpp
-    weightedModuleVectorFuncs.cpp
-    widgetFuncs.cpp
-    widgetGroupHeaderFuncs.cpp
-    base.h
-    vec3Labels.h
-    producers.h
-)
-
 # Target 'keywordWidgets'
-add_library(keywordWidgets ${keywordWidgets_UIS_H} ${keywordWidgets_SRCS} ${keywordWidgets_MOC_SRCS})
+add_library(
+  keywordWidgets
+  atomTypeVector.cpp
+  atomTypeVector.h
+  atomTypeVector.ui
+  base.cpp
+  base.h
+  bool.cpp
+  bool.h
+  CMakeLists.txt
+  configuration.cpp
+  configuration.h
+  configuration.ui
+  configurationVector.cpp
+  configurationVector.h
+  configurationVector.ui
+  dialog.cpp
+  dialog.h
+  dialog.ui
+  double.cpp
+  double.h
+  dropDown.cpp
+  dropDown.h
+  dropDown.ui
+  dropWidget.cpp
+  dropWidget.h
+  enumOptions.cpp
+  enumOptions.h
+  expressionVariableVector.cpp
+  expressionVariableVector.h
+  expressionVariableVector.ui
+  fileAndFormat.cpp
+  fileAndFormat.h
+  fileAndFormat.ui
+  function1D.cpp
+  function1D.h
+  function1D.ui
+  integer.cpp
+  integer.h
+  interactionPotential.cpp
+  interactionPotential.h
+  interactionPotential.ui
+  isotopologueSet.cpp
+  isotopologueSet.h
+  isotopologueSet.ui
+  layer.cpp
+  layer.h
+  layer.ui
+  module.cpp
+  module.h
+  module.ui
+  moduleVector.cpp
+  moduleVector.h
+  moduleVector.ui
+  nodeAndInteger.cpp
+  nodeAndInteger.h
+  nodeAndInteger.ui
+  node.cpp
+  node.h
+  node.ui
+  nodeValue.cpp
+  nodeValueEnumOptions.cpp
+  nodeValueEnumOptions.h
+  nodeValueEnumOptions.ui
+  nodeValue.h
+  nodeValue.ui
+  nodeVector.cpp
+  nodeVector.h
+  nodeVector.ui
+  optionalDouble.cpp
+  optionalDouble.h
+  optionalInt.cpp
+  optionalInt.h
+  producers.cpp
+  producers.h
+  range.cpp
+  range.h
+  range.ui
+  rangeVector.cpp
+  rangeVector.h
+  rangeVector.ui
+  species.cpp
+  species.h
+  speciesSite.cpp
+  speciesSite.h
+  speciesSite.ui
+  speciesSiteVector.cpp
+  speciesSiteVector.h
+  speciesSiteVector.ui
+  species.ui
+  speciesVector.cpp
+  speciesVector.h
+  speciesVector.ui
+  stdString.cpp
+  stdString.h
+  vec3Double.cpp
+  vec3Double.h
+  vec3Double.ui
+  vec3Integer.cpp
+  vec3Integer.h
+  vec3Integer.ui
+  vec3Labels.cpp
+  vec3Labels.h
+  vec3NodeValue.cpp
+  vec3NodeValue.h
+  vec3NodeValue.ui
+  weightedModuleVector.cpp
+  weightedModuleVector.h
+  weightedModuleVector.ui
+  widget.cpp
+  widgetGroupHeader.cpp
+  widgetGroupHeader.h
+  widgetGroupHeader.ui
+  widget.h
+)
 target_include_directories(keywordWidgets PRIVATE ${PROJECT_SOURCE_DIR}/src ${PROJECT_BINARY_DIR}/src ${Qt6Widgets_INCLUDE_DIRS})
 target_link_libraries(keywordWidgets PRIVATE base)
