+++ src/gui/keywordWidgets/atomTypeVector.cpp
