+++ src/gui/keywordWidgets/bool.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2024 Team Dissolve and contributors
 
-#include "gui/keywordWidgets/bool.hui"
+#include "gui/keywordWidgets/bool.h"
 
 BoolKeywordWidget::BoolKeywordWidget(QWidget *parent, BoolKeyword *keyword, const CoreData &coreData)
     : QCheckBox(parent), KeywordWidgetBase(coreData), keyword_(keyword)
