+++ src/gui/keywordWidgets/configuration.cpp
