+++ src/gui/keywordWidgets/configurationVector.cpp
