+++ src/gui/keywordWidgets/dropWidget.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2024 Team Dissolve and contributors
 
-#include "gui/keywordWidgets/dropWidget.hui"
+#include "gui/keywordWidgets/dropWidget.h"
 #include <stdio.h>
 
 DropWidget::DropWidget(QWidget *parent) : QWidget(parent, Qt::FramelessWindowHint | Qt::Popup) {}
