+++ src/gui/keywordWidgets/enumOptions.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2024 Team Dissolve and contributors
 
 #include "gui/helpers/mouseWheelAdjustmentGuard.h"
-#include "gui/keywordWidgets/enumOptions.hui"
+#include "gui/keywordWidgets/enumOptions.h"
 
 EnumOptionsKeywordWidget::EnumOptionsKeywordWidget(QWidget *parent, EnumOptionsBaseKeyword *keyword, const CoreData &coreData)
     : QComboBox(parent), KeywordWidgetBase(coreData), keyword_(keyword)
