+++ src/gui/keywordWidgets/expressionVariableVector.cpp
