+++ src/gui/keywordWidgets/interactionPotential.cpp
