+++ src/gui/keywordWidgets/module.cpp
