+++ src/gui/keywordWidgets/moduleVector.cpp
