+++ src/gui/keywordWidgets/nodeAndInteger.cpp
