+++ src/gui/keywordWidgets/nodeValue.cpp
