+++ src/gui/keywordWidgets/nodeVector.cpp
