+++ src/gui/keywordWidgets/stdString.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2024 Team Dissolve and contributors
 
-#include "gui/keywordWidgets/stdString.hui"
+#include "gui/keywordWidgets/stdString.h"
 
 StringKeywordWidget::StringKeywordWidget(QWidget *parent, StringKeyword *keyword, const CoreData &coreData)
     : QLineEdit(parent), KeywordWidgetBase(coreData), keyword_(keyword)
