+++ src/gui/keywordWidgets/vec3NodeValue.cpp
