+++ src/gui/keywordWidgets/weightedModuleVector.cpp
