+++ src/gui/keywordWidgets/widgetGroupHeader.cpp
